fix: tap targeting
LLM APIs deals with pixels, but the tap tool deals with points. These differ in
ways both predictable (retina screen scaling factors) and not (different LLM
APIs have different max dimensions for images, and different MCP clients could
use different LLM APIs). As a result: tap accuracy was atrocious in my usage of
this with claude code.

This patch addresses the issue by fetching the screen dimensions in points at
the time of screenshots, and scaling the returned image to fit. That way, points
are 1:1 with the pixels the LLM sees, and they have no trouble tapping.